% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhs.R
\name{rhs}
\alias{rhs}
\title{Rijkshuisstijlkleuren}
\usage{
rhs(...)
}
\arguments{
\item{...}{Komma gescheiden character met kleurnamen uit het Rijkshuisstijlpalet}
}
\value{
Character vector met kleurdefinities
}
\description{
Rijkshuisstijlkleuren
}
\examples{
rhs("paars", "groen")
}
