% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhs.R
\name{rhs}
\alias{rhs}
\title{Rijkshuisstijlkleuren}
\usage{
rhs(col)
}
\arguments{
\item{col}{Character vector met kleurnamen uit het Rijkshuisstijlpalet. Hoofdletterongevoelig.}
}
\value{
Character vector met kleurdefinities.
}
\description{
Rijkshuisstijlkleuren
}
\details{
Geldige kleurnamen zijn:

Paars, Violet, Robijnrood, Roze, Rood, Oranje, Donkergeel, Geel, Donkerbruin, Bruin,
Donkergroen, Groen, Mosgroen,  Mintgroen, Donkerblauw, Hemelblauw, Lichtblauw

Ieder kleur heeft ook twee lichtere tinten door ze aan te vullen met 1 of 2.

Grijs 1 t/m 7 zijn neutrale tinten voor achtergrondkaders, grafieken en tabellen.
}
\examples{
rhs(col = c("paars", "groen"))

}
\references{
\url{https://www.rijkshuisstijl.nl/basiselementen/basiselementen-online/online-kleuren}
}
